+++ CurrencySpender/Configuration/Config.cs
@@ -9,6 +9,7 @@ public class Config: IEzConfig
     public string Version { get; set; } = "0.0.0.0";
 
     public List<TrackedCurrency> Currencies = [];
+    public int Seperator = 0;
     //public Dictionary<uint, uint> FateRanks = [];
     //public List<BuyableItem> Items = [];