+++ CurrencySpender/Windows/DebugMainTab.cs
@@ -22,70 +22,17 @@ internal class DebugMainTab
             //ImGui.TableSetupColumn("Price");
             //ImGui.TableSetupColumn("");
             ImGui.TableHeadersRow();
-            foreach (var item in Generator.items.Where(item => item.Shop.NpcId == 1037055).ToList())
+            foreach (var shop in Generator.shops.ToList())
             {
                 ImGui.TableNextColumn();
                 //ImGui.Text("text1");
-                ImGui.Text(item.ShopId.ToString());
+                ImGui.Text($"{shop.ShopId}");
                 ImGui.TableNextColumn();
-                ImGui.Text(item.Shop.RequiredLevel.ToString());
-                ImGui.TableNextColumn();
-                //ImGui.Text("text2");
-                ImGui.Text(item.Name);
-                //ImGui.TableNextColumn();
-                //ImGui.Text("text3");
-                //ImGui.Text(shop.NpcId.ToString());
-                //ImGui.TableNextColumn();
-                //ImGui.Text("text4");
-                //ImGui.Text(shop.NpcName);
-                //ImGui.TableNextColumn();
-                //ImGui.Text(shop.Location.NpcName.ToString());
-                //ImGui.TableNextColumn();
-                //ImGui.Text(item.ENpcData[0].fields.ToString());
-            }
-            ImGui.TableNextColumn();
-            ImGui.TableNextColumn();
-            foreach (var item in Generator.items.Where(item => item.ShopId == 1770754).ToList())
-            {
-                ImGui.TableNextColumn();
-                //ImGui.Text("text1");
-                ImGui.Text(item.ShopId.ToString());
+                ImGui.Text($"{shop.NpcId}");
                 ImGui.TableNextColumn();
                 //ImGui.Text("text2");
-                ImGui.Text(item.Name);
-                //ImGui.TableNextColumn();
-                //ImGui.Text("text3");
-                //ImGui.Text(shop.NpcId.ToString());
-                //ImGui.TableNextColumn();
-                //ImGui.Text("text4");
-                //ImGui.Text(shop.NpcName);
-                //ImGui.TableNextColumn();
-                //ImGui.Text(shop.Location.NpcName.ToString());
-                //ImGui.TableNextColumn();
-                //ImGui.Text(item.ENpcData[0].fields.ToString());
+                ImGui.Text($"{shop.NpcName}");
             }
-            ImGui.TableNextColumn();
-            ImGui.TableNextColumn();
-            foreach (var item in Generator.items.Where(item => item.ShopId == 1770755).ToList())
-            {
-                ImGui.TableNextColumn();
-                //ImGui.Text("text1");
-                ImGui.Text(item.ShopId.ToString());
-                ImGui.TableNextColumn();
-                //ImGui.Text("text2");
-                ImGui.Text(item.Name);
-                //ImGui.TableNextColumn();
-                //ImGui.Text("text3");
-                //ImGui.Text(shop.NpcId.ToString());
-                //ImGui.TableNextColumn();
-                //ImGui.Text("text4");
-                //ImGui.Text(shop.NpcName);
-                //ImGui.TableNextColumn();
-                //ImGui.Text(shop.Location.NpcName.ToString());
-                //ImGui.TableNextColumn();
-                //ImGui.Text(item.ENpcData[0].fields.ToString());
-            }
-            //}
             //PluginLog.Verbose("Starting ImGui EndTable rendering...");
             ImGui.EndTable();
         }